Move from association_service to role_service
Role methods are admin related functions. Only users with role admin
can access said service, so it's a better fit in admin service.
