+++ src/controller/admin_controller.rs
@@ -10,13 +10,16 @@ use crate::{
     model::session_model::CurrentUser,
     service::{
         admin_service::{get_roles_service, get_stats_of_tables_service},
-        association_service::{add_rol_to_user_service, delete_rol_to_user_service},
         baby_service::{get_all_babies_service, get_baby_by_id_service},
+        role_service::{
+            add_rol_to_user_service, delete_rol_to_user_service, get_role_by_name_service,
+        },
         session_service::current_user_is_admin,
         user_service::{
-            delete_active_user_service, delete_old_users_service, delete_user_with_time_constrain_service,
-            get_all_users_service, get_user_id_from_username,
-        }, role_service::get_role_by_name_service,
+            delete_active_user_service, delete_old_users_service,
+            delete_user_with_time_constrain_service, get_all_users_service,
+            get_user_id_from_username,
+        },
     },
 };
