+++ src/service/mod.rs
@@ -1,12 +1,11 @@
-pub mod association_service;
+pub mod admin_service;
 pub mod baby_service;
 pub mod dream_service;
 pub mod dream_summary_service;
 pub mod meal_service;
 pub mod meal_summary_service;
+pub mod role_service;
 pub mod session_service;
 pub mod user_service;
 pub mod util_service;
 pub mod weight_service;
-pub mod admin_service;
-pub mod role_service;
